+++ plotnine/stats/stat_count.py
@@ -51,7 +51,7 @@ class stat_count(stat):
             msg = 'stat_count() must not be used with a y aesthetic'
             raise PlotnineError(msg)
 
-        weight = data.get('weight', np.ones(len(x)))
+        weight = data.get('weight', np.ones(len(x), dtype=int))
         width = params['width']
         if width is None:
             width = resolution(x, False) * 0.9
@@ -61,7 +61,7 @@ class stat_count(stat):
             'weight', index=['x'], aggfunc=np.sum)['weight']
         x = count.index
         count = count.values
-        return pd.DataFrame({'count': count.astype(int),
+        return pd.DataFrame({'count': count,
                              'prop': count / np.abs(count).sum(),
                              'x': x,
                              'width': width})